IN PROGRESS: work re #5, #7, #10, #11, #15, #16
