+++ ebagis/serializers/upload.py
@@ -1,52 +1,43 @@
 from __future__ import absolute_import
-
 import os
+
+from django.contrib.contenttypes.models import ContentType
 
 from rest_framework import serializers
 
 from ..models.aoi import AOI
-from ..models.upload import AOIUpload, UpdateUpload
+from ..models.upload import Upload
 
-from ..utilities import validate_path
+from ..utils.validation import validate_path
 
-from .task import AOITaskSerializer
+from .task import TaskSerializer
 
 
-class AOIUploadSerializer(serializers.ModelSerializer):
+class UploadSerializer(serializers.ModelSerializer):
     url = serializers.HyperlinkedIdentityField(view_name="aoiupload-detail")
     user = serializers.HyperlinkedRelatedField(view_name="user-detail",
                                                read_only=True)
     task = AOITaskSerializer(read_only=True)
     md5 = serializers.CharField(required=False)
 
-    def validate_filename(self, value):
-        aoi_name = os.path.splitext(value)[0]
-        validate_path(aoi_name, allow_whitespace=True)
-
-        try:
-            AOI.objects.get(name=aoi_name)
-        except AOI.DoesNotExist:
-            pass
-        else:
-            raise Exception("An AOI of the same name already exists.")
-
-        return value
-
-    class Meta:
-        model = AOIUpload
-        read_only_fields = ('status', 'completed_at', 'task', 'offset')
-
-
-class UpdateUploadSerializer(serializers.ModelSerializer):
-    url = serializers.HyperlinkedIdentityField(view_name="updateupload-detail")
-    user = serializers.HyperlinkedRelatedField(view_name="user-detail",
-                                               read_only=True)
-
-    def validate_filename(self, value):
-        file_name = os.path.splitext(value)
-        validate_path(file_name)
-        return value
+    #def validate_filename(self, value):
+    #    name = os.path.splitext(value)[0]
+    #    validate_path(name, allow_whitespace=True)
+    #    upload_class = ContentType.model_class(upload.content_type)
+    #
+    #    try:
+    #        upload_class.objects.get(name)
+    #    except upload_class.DoesNotExist:
+    #        pass
+    #    else:
+    #        raise Exception("An object of the same name and type already exists.")
+    #
+    #    return value
 
     class Meta:
-        model = UpdateUpload
-        read_only_fields = ('status', 'completed_at', 'processed')
+        model = Upload
+        read_only_fields = (
+            'status', 'completed_at', 'task', 'offset', 'content_type',
+            'object_id', 'update',
+        )
+
